バリデーション

+++ resources/views/post/show.blade.php
@@ -1,9 +1,10 @@
-@extends('layouts.layout2')
+@extends('layouts.layout')
 
 @section('content')
 <div class="title_box">
     <p class="post_title">{{ $post->post }}</p>
     <div class="edit_delete_box">
+    @if($post->user->name === auth()->user()->name )
         <div class="edit_button">
             <p><a href="/themes/{{ $theme->id }}/posts/{{ $post->id }}/edit"><i class="far fa-edit"></i>edit</a></p>
         </div>
@@ -14,6 +15,7 @@
                 <span class="delete_link" onclick="return deletePost(this);"><i class="far fa-trash-alt"></i><input type="submit" style="display:none">delete</span>
             </form>
         </div>
+    @endif
         <div class="like_button">
             @if($post->is_liked_by_auth_user())
                 <a href="{{ route('post.unlike',['id' => $post->id])  }}" class="btn">いいね<span class="badge">{{ $post->likes->count() }}</span></a>
@@ -29,27 +31,31 @@
     </div>
 </div>
     <section class="comments">
-        <p class="heading">コメント</p>
+        <p class="heading">コメント<span class="limit_letters">(最大200字)</span></p>
         <div class="content">
             <form class="comment_form" action="/themes/{{ $theme->id }}/posts/{{ $post->id }}/comments" method="post">
                 {{ csrf_field() }}
-                <input type="text" name="comment[body]" placeholder="コメントしてみてね(匿名)"/>
-                <input type="submit" value="store"/>
+                <input type="text" name="comment[body]" placeholder="コメントしてみてね(匿名)" value="{{ old('comment.body') }}"/>
+                <p class="error_message">{{ $errors->first('comment.body') }}</p>
+                <input type="submit" class="submit_button" value="送信"/>
             </form>
             <div class="comment_list">
                 @if (!$post->comments->isEmpty())
                     @foreach ($comments as $comment)
                         <div class="comment">
                             <p class="comment_text">{{ $comment->body }}</p>
-                            <div class="edit_button">
-                                <p><a href="/themes/{{ $theme->id }}/posts/{{ $post->id }}/comments/{{ $comment->id }}/edit"><i class="far fa-edit"></i>edit</a></p>
-                            </div>
-                            <div class="delete_button">
-                                <form action="/themes/{{ $theme->id }}/posts/{{ $post->id }}/comments/{{ $comment->id }}" id="form_delete_comment" method="post">
-                                    {{ csrf_field() }}
-                                    {{ method_field('delete') }}
-                                    <span class="delete_link" onclick="return deleteComment(this);"><i class="far fa-trash-alt"></i><input type="submit" style="display:none">delete</span>
-                                </form>
+                            <div class="comment_detail">
+                                <p>by {{ optional($comment->user)->name }}</p>
+                                @if($comment->user->name === auth()->user()->name )
+                                    <div class="comment_edit_delete">
+                                        <p><a href="/themes/{{ $theme->id }}/posts/{{ $post->id }}/comments/{{ $comment->id }}/edit"><i class="far fa-edit"></i>edit</a></p>
+                                        <form action="/themes/{{ $theme->id }}/posts/{{ $post->id }}/comments/{{ $comment->id }}" id="form_delete_comment" method="post">
+                                            {{ csrf_field() }}
+                                            {{ method_field('delete') }}
+                                            <span class="delete_link" onclick="return deleteComment(this);"><i class="far fa-trash-alt"></i><input type="submit" style="display:none">delete</span>
+                                        </form>
+                                    </div>
+                                @endif
                             </div>
                         </div>
                     @endforeach
