+++ resources/views/theme/show.blade.php
@@ -5,18 +5,20 @@
     <div class="theme_title">
         <p>{{ $theme->theme }}</p>
     </div>
-    <div class="edit_delete_box">
-        <div class="edit_button">
-            <p><a href="/themes/{{ $theme->id }}/edit"><i class="far fa-edit"></i>edit</a></p>
+    @if($theme->user->name === auth()->user()->name )
+        <div class="edit_delete_box">
+            <div class="edit_button">
+                <p><a href="/themes/{{ $theme->id }}/edit"><i class="far fa-edit"></i>edit</a></p>
+            </div>
+            <div class="delete_button">
+                <form action="/themes/{{ $theme->id }}" id="form_delete" method="post">
+                    {{ csrf_field() }}
+                    {{ method_field('delete') }}
+                    <span class="delete_link" onclick="return deleteTheme(this);"><i class="far fa-trash-alt"></i><input type="submit" style="display:none">delete</span>
+                </form>
+            </div>
         </div>
-        <div class="delete_button">
-            <form action="/themes/{{ $theme->id }}" id="form_delete" method="post">
-                {{ csrf_field() }}
-                {{ method_field('delete') }}
-                <span class="delete_link" onclick="return deleteTheme(this);"><i class="far fa-trash-alt"></i><input type="submit" style="display:none">delete</span>
-            </form>
-        </div>
-    </div>
+    @endif
     <div class="theme_explain">
         <p class="theme_detail">{{ date('Y/m/d',strtotime($theme->created_at)) }}</p>
         <p class="theme_detail">by {{ optional($theme->user)->name }}</p>
@@ -31,11 +33,12 @@
     </div>
 </div>
 <section class="new_post">
-    <p class="heading">回答する</p>
+    <p class="heading">回答する<span class="limit_letters">(最大200字)</span></p>
     <div class="content">
         <form action="/themes/{{ $theme->id }}/posts" method="post">
             {{ csrf_field() }}
-            <input type="text" name="post[post]" placeholder="テーマを投稿してね"/>
+            <input type="text" name="post[post]" placeholder="テーマを投稿してね" value="{{ old('post.post') }}"/>
+            <p class="error_message">{{ $errors->first('post.post') }}</p>
             <input type="submit" value="store"/>
         </form>
     </div>
